+++ chaosaws/asg/actions.py
@@ -13,7 +13,7 @@ from chaoslib.types import Configuration, Secrets
 from logzero import logger
 
 __all__ = ["suspend_processes", "resume_processes",
-           "terminate_random_instances"]
+           "terminate_random_instances", "detach_random_instances"]
 
 
 def terminate_random_instances(asg_names: List[str] = None,
@@ -197,6 +197,73 @@ def resume_processes(asg_names: List[str] = None,
 
     return get_asg_by_name(
         [a['AutoScalingGroupName'] for a in asgs['AutoScalingGroups']], client)
+
+
+def detach_random_instances(asg_names: List[str] = None,
+                            tags: List[dict] = None,
+                            instance_count: int = None,
+                            instance_percent: int = None,
+                            decrement_capacity: bool = False,
+                            configuration: Configuration = None,
+                            secrets: Secrets = None) -> AWSResponse:
+    """
+    Detaches one or more random instances from an autoscaling group
+
+    Parameters:
+        One of:
+            asg_names: a list of one or more asg names
+            tags: a list of key/value pair to identify asg(s) by
+
+        One of:
+            instance_count: integer value of number of instances to detach
+            instance_percent: 1-100, percent of instances to detach
+
+        decrement_capacity: boolean value to determine if the desired capacity
+        of the autoscaling group should be decreased
+
+    `tags` are expected as a list of dictionary objects:
+    [
+        {'Key': 'TagKey1', 'Value': 'TagValue1'},
+        {'Key': 'TagKey2', 'Value': 'TagValue2'},
+        ...
+    ]
+    """
+    validate_asgs(asg_names, tags)
+
+    if not any([instance_count, instance_percent]):
+        raise FailedActivity('You must specify either "instance_count" or '
+                             '"instance_percent"')
+
+    client = aws_client('autoscaling', configuration, secrets)
+
+    if asg_names:
+        asgs = get_asg_by_name(asg_names, client)
+    else:
+        asgs = get_asg_by_tags(tags, client)
+
+    results = {}
+    for a in asgs['AutoScalingGroups']:
+        # Filter out all instances not currently 'InService'
+        instances = [e['InstanceId'] for e in a['Instances'] if e[
+            'LifecycleState'] == 'InService']
+
+        if instance_percent:
+            instance_count = int(float(
+                len(instances) * float(instance_percent)) / 100)
+
+        if instance_count > len(instances):
+            raise FailedActivity('You are attempting to detach more instances '
+                                 'than exist on asg %s' % (
+                                    a['AutoScalingGroupName']))
+
+        instances = random.sample(instances, instance_count)
+
+        response = client.detach_instances(
+            AutoScalingGroupName=a['AutoScalingGroupName'],
+            InstanceIds=sorted(instances),
+            ShouldDecrementDesiredCapacity=decrement_capacity)
+        results.setdefault('Activities', []).extend(response['Activities'])
+    return results
 
 
 ###############################################################################